requirements.txt
opencv-python==4.12.0.88
numpy==2.2.6
watchdog==6.0.0
python-dotenv==1.1.1
ffmpeg-python==0.2.0
ultralytics==8.3.179
